Fixed some SDK build warnings.

+++ sdk/esp-open-rtos-rsf/core/phy_info.c
@@ -150,7 +150,7 @@ void dump_phy_info(const sdk_phy_info_t *info, bool raw)
     if(raw) {
         printf("Raw values:");
         uint8_t *p = (uint8_t *)info;
-        for(int i = 0; i < sizeof(sdk_phy_info_t); i ++) {
+        for(size_t i = 0; i < sizeof(sdk_phy_info_t); i ++) {
             if(i % 8 == 0) {
                 printf("\n0x%02x:", i);
             }
